Fixed Windows compile errors
When FIt-SNE was originally adapted to Windows, key changes were made to
annoylib.h, mman.c, and mman.h in order to allow it to compile on
Windows. When we updated annoylib.h, those changes were not carried
forward. This commit restores those changes to the new annoylib.h file,
so that it can be compiled on Windows.